requirements.txt
mcp
pandas
fubon_neo-2.2.0-cp37-abi3-macosx_11_0_arm64.whl
pydantic>=2.10.0
